Move print_utils.h.
